server/requirements.txt
uv
starlette
python-dotenv>=0.21.0
requests==2.32.3
flask-cors 
google-generativeai
Werkzeug==2.0.1
gunicorn==20.1.0
langchain==0.3.0
langchain_community==0.3.0
langchain_openai==0.2.0
langchain_core==0.3.0
langgraph==0.2.21
sentence-transformers
chromadb
